arrange into packages, rename Info to Subject
